requirements.txt
fastapi
uvicorn
pillow
scikit-learn
opencv-python
tensorflow
xgboost
numpy
pandas
python-multipart
joblib
